+++ src/commands/Dashboard.ts
@@ -270,6 +270,14 @@ export class Dashboard {
       }`
     ];
 
+    const frontMatterUri = webView
+      .asWebviewUri(
+        SettingsHelper.projectConfigPath ? Uri.file(SettingsHelper.projectConfigPath) : Uri.file('')
+      )
+      .toString();
+
+    const webviewUrl = frontMatterUri.replace(`/${SettingsHelper.globalFile}`, '');
+
     return `
       <!DOCTYPE html>
       <html lang="en" style="width:100%;height:100%;margin:0;padding:0;">
@@ -285,7 +293,9 @@ export class Dashboard {
       isBeta ? 'BETA' : 'main'
     }" data-version="${version.usedVersion}" style="width:100%;height:100%;margin:0;padding:0;" ${
       version.usedVersion ? '' : `data-showWelcome="true"`
-    } ${experimental ? `data-experimental="${experimental}"` : ''} ></div>
+    } ${
+      experimental ? `data-experimental="${experimental}"` : ''
+    } data-webview-url="${webviewUrl}" ></div>
 
         ${(scriptsToLoad || [])
           .map((script) => {